+++ packages/box-core/changelog.md
@@ -9,3 +9,4 @@
 - Rename extendControls in registerComponents
 - Rename _fields in _children
 - Rename renderFields in renderChildren
+- Rename _fromId in _id